Rename limits.h to limits.hpp
