+++ source/limits.hpp
